brought logic into DestroyWindow.cs
