+++ Assets/Resources/UI/PopUp/GardenPopUp.prefab
@@ -611,15 +611,12 @@ MonoBehaviour:
     _scalingTime: 0.25
     _scalingEase: 1
   _closeButton: {fileID: 7139061424278485610}
-  _destroyButton: {fileID: 8623747221757070152}
   _openDestroyWindowButton: {fileID: 3651266083829886753}
   _closeDestroyWindowButton: {fileID: 1509832054607647997}
-  _cancelBatton: {fileID: 8639576635962316806}
   _seedsWindow: {fileID: 57441253066640029}
   _destroyWindow: {fileID: 604140775829921635}
   _cards: {fileID: 8787328670163120845}
   _tab: {fileID: 3922789393403875680}
-  _destroyEffect: {fileID: 6226165539581104974, guid: 24efb91e65d324d44aa2cd4f00e9aa22, type: 3}
   _slot: {fileID: 8493690943087483406}
 --- !u!1 &1262613149714536942
 GameObject:
@@ -2231,7 +2228,7 @@ GameObject:
   - component: {fileID: 8352650075511445207}
   - component: {fileID: 8623747221757070152}
   m_Layer: 0
-  m_Name: ButtonAprove
+  m_Name: ButtonApprove
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -2374,7 +2371,7 @@ RectTransform:
   m_AnchorMin: {x: 0.31625003, y: 0.38000003}
   m_AnchorMax: {x: 0.68375003, y: 0.79}
   m_AnchoredPosition: {x: 0, y: 8}
-  m_SizeDelta: {x: 0, y: 0}
+  m_SizeDelta: {x: 20, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &3553094271434087342
 CanvasRenderer:
@@ -2613,7 +2610,10 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d33c227c5f55b9c49a4218ec6cd7ce2d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _scalingTime: 0
+  _scalingTime: 0.1
+  _destroyButton: {fileID: 8623747221757070152}
+  _cancelBatton: {fileID: 8639576635962316806}
+  _destroyEffect: {fileID: 6226165539581104974, guid: 24efb91e65d324d44aa2cd4f00e9aa22, type: 3}
 --- !u!1 &5829934011595294867
 GameObject:
   m_ObjectHideFlags: 0
